+++ src/main/java/com/Airline/dao/entity/BookingRecord.java
@@ -14,7 +14,7 @@ public class BookingRecord {
 
 	@Id
 	@Column(name="booking_id")
-	private Integer bookingId;
+	private Long bookingId;
 	
 	@Column(name="booking_date")
 	private Date bookingDate;
@@ -44,11 +44,11 @@ public class BookingRecord {
 		
 	}
 
-	public Integer getBookingId() {
+	public Long getBookingId() {
 		return bookingId;
 	}
 
-	public void setBookingId(Integer bookingId) {
+	public void setBookingId(Long bookingId) {
 		this.bookingId = bookingId;
 	}
